Fixed missing typehints, some code refactoring regarding AbstractTaxHandler
