+++ src/Adapter/Tax/AbstractTaxHandler.php
@@ -24,7 +24,7 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShop\PrestaShop\Adapter\Tax\CommandHandler;
+namespace PrestaShop\PrestaShop\Adapter\Tax;
 
 use PrestaShop\PrestaShop\Core\Domain\Tax\Exception\TaxNotFoundException;
 use PrestaShop\PrestaShop\Core\Domain\Tax\ValueObject\TaxId;
